🐛 (macros): Prevent faulty members at compile time
When generating trait method names from XML member names,
successive capital letters later on lead to faulty recreated
member names.

This commit allows uses to define a path to the XML file as
trait attribute.

```rust
    interface = "org.a11y.atspi.Accessible",
    xml_path = "../xml/Accessible.xml"
)]
```

If any of the method names' corresponding member is not
found in the XML definitions, it will now suggest to decorate
the item appropriately to fix the problem.